Refactor CLI commands into separate modules
